SI-7848 Xlint says what looks interpolated
The motivating use case was an Expecty debug string
getting flagged for `$eq`.

The test case demonstrates a different bug, in which
the position of the literal tree is changed when typer
gets rid of the unused local, so that when the tree is
re-typed in erasure, a second, spurious warning is
emitted at the start of the method. Specifically, the
second warning is not suppressed because of the
different position.