Position#show prints the point.
Can't ensure range position points are meaningful when we never
see them. To limit noise, only print the point when it != start.

  [x:y]    // point=x, start=x, end=y
  [p/x:y]  // point=p, start=x, end=y

I'm open to a different syntax.

Also prints NoPosition as [X] rather than [NoPosition] because
noise is for construction workers and attenders of rock concerts.

Some range position and parser tests are included so we can see
the checkfile change when the forthcoming fix happens (either an
error message improvement or a positional one.)